improved vector detector
